FAISS bug fix

+++ benchmarks/early_termination_faiss.cpp
@@ -29,9 +29,9 @@ int main(int argc, char* argv[]) {
         return 1;
     }
 
-    const int n = it->second.first;
-    const int d = it->second.second;
-    const int n_clusters =
+    const size_t n = it->second.first;
+    const size_t d = it->second.second;
+    const size_t n_clusters =
         std::max<int>(1u, static_cast<int>(std::sqrt(static_cast<double>(n)) * 4.0));
     const size_t THREADS = omp_get_max_threads();
     omp_set_num_threads(THREADS);
