+++ benchmarks/end_to_end_faiss.cpp
@@ -31,9 +31,9 @@ int main(int argc, char* argv[]) {
         return 1;
     }
 
-    const int n = it->second.first;
-    const int d = it->second.second;
-    const int n_clusters =
+    const size_t n = it->second.first;
+    const size_t d = it->second.second;
+    const size_t n_clusters =
         std::max<int>(1u, static_cast<int>(std::sqrt(static_cast<double>(n)) * 4.0));
     int n_iters = bench_utils::MAX_ITERS;
     const size_t THREADS = omp_get_max_threads();
